+++ server/routers/user/deleteUser.ts
@@ -6,6 +6,7 @@ import { eq } from 'drizzle-orm';
 import response from "@server/utils/response";
 import HttpCode from '@server/types/HttpCode';
 import createHttpError from 'http-errors';
+import { ActionsEnum, checkUserActionPermission } from '@server/auth/actions';
 
 const deleteUserSchema = z.object({
   userId: z.string().uuid()
@@ -24,6 +25,12 @@ export async function deleteUser(req: Request, res: Response, next: NextFunction
     }
 
     const { userId } = parsedParams.data;
+
+    // Check if the user has permission to list sites
+    const hasPermission = await checkUserActionPermission(ActionsEnum.deleteUser, req);
+    if (!hasPermission) {
+      return next(createHttpError(HttpCode.FORBIDDEN, 'User does not have permission to list sites'));
+    }
 
     const deletedUser = await db.delete(users)
       .where(eq(users.id, userId))
